Flyttar brevdata-uthenting-kode, vil også her ha mest mogleg lik løype, sjølv om resultatet per no er veldig mykje enklare for infobrev
