+++ apps/etterlatte-brev-api/src/main/kotlin/no/nav/etterlatte/brev/PDFGenerator.kt
@@ -10,10 +10,17 @@ import no.nav.etterlatte.brev.db.BrevRepository
 import no.nav.etterlatte.brev.hentinformasjon.BrevdataFacade
 import no.nav.etterlatte.brev.model.Brev
 import no.nav.etterlatte.brev.model.BrevData
+import no.nav.etterlatte.brev.model.BrevDataMapper
 import no.nav.etterlatte.brev.model.BrevDataMapper.BrevkodePar
 import no.nav.etterlatte.brev.model.BrevID
+import no.nav.etterlatte.brev.model.BrevProsessType
+import no.nav.etterlatte.brev.model.InnholdMedVedlegg
+import no.nav.etterlatte.brev.model.ManueltBrevData
+import no.nav.etterlatte.brev.model.ManueltBrevMedTittelData
 import no.nav.etterlatte.brev.model.Pdf
+import no.nav.etterlatte.brev.model.bp.OmregnetBPNyttRegelverk
 import no.nav.etterlatte.brev.model.bp.OmregnetBPNyttRegelverkFerdig
+import no.nav.etterlatte.libs.common.Vedtaksloesning
 import no.nav.etterlatte.libs.common.retryOgPakkUt
 import no.nav.etterlatte.token.BrukerTokenInfo
 import org.slf4j.LoggerFactory
@@ -21,8 +28,10 @@ import org.slf4j.LoggerFactory
 class PDFGenerator(
     private val db: BrevRepository,
     private val brevDataFacade: BrevdataFacade,
+    private val brevDataMapper: BrevDataMapper,
     private val adresseService: AdresseService,
     private val brevbakerService: BrevbakerService,
+    private val migreringBrevDataService: MigreringBrevDataService,
 ) {
     private val logger = LoggerFactory.getLogger(this::class.java)
 
@@ -32,7 +41,6 @@ class PDFGenerator(
         automatiskMigreringRequest: MigreringBrevRequest?,
         avsenderRequest: (GenerellBrevData) -> AvsenderRequest,
         brevKode: (GenerellBrevData, Brev, MigreringBrevRequest?) -> BrevkodePar,
-        brevData: suspend (BrevDataRequest) -> BrevData,
         lagrePdfHvisVedtakFattet: (GenerellBrevData, Brev, Pdf) -> Unit = { _, _, _ -> run {} },
     ): Pdf {
         val brev = db.hentBrev(id)
@@ -51,13 +59,11 @@ class PDFGenerator(
         val sak = generellBrevData.sak
         val letterData =
             brevData(
-                BrevDataRequest(
-                    generellBrevData,
-                    automatiskMigreringRequest,
-                    brev,
-                    bruker,
-                    brevkodePar,
-                ),
+                generellBrevData,
+                automatiskMigreringRequest,
+                brev,
+                bruker,
+                brevkodePar,
             )
         val brevRequest =
             BrevbakerRequest.fra(
@@ -93,12 +99,69 @@ class PDFGenerator(
             }
         }.also { lagrePdfHvisVedtakFattet(generellBrevData, brev, it) }
     }
-}
 
-data class BrevDataRequest(
-    val generellBrevData: GenerellBrevData,
-    val automatiskMigreringRequest: MigreringBrevRequest? = null,
-    val brev: Brev,
-    val bruker: BrukerTokenInfo,
-    val brevkodePar: BrevkodePar,
-)
+    private suspend fun brevData(
+        generellBrevData: GenerellBrevData,
+        automatiskMigreringRequest: MigreringBrevRequest?,
+        brev: Brev,
+        brukerTokenInfo: BrukerTokenInfo,
+        brevkodePar: BrevkodePar,
+    ): BrevData {
+        if (brevkodePar.ferdigstilling == EtterlatteBrevKode.TOM_MAL_INFORMASJONSBREV) {
+            return ManueltBrevMedTittelData(
+                requireNotNull(db.hentBrevPayload(brev.id)).elements,
+                brev.tittel,
+            )
+        }
+        return when (
+            generellBrevData.systemkilde == Vedtaksloesning.PESYS ||
+                automatiskMigreringRequest?.erOmregningGjenny ?: false
+        ) {
+            false -> opprettBrevData(brev, generellBrevData, brukerTokenInfo, brevkodePar)
+            true ->
+                OmregnetBPNyttRegelverkFerdig(
+                    innhold =
+                        InnholdMedVedlegg(
+                            { hentLagretInnhold(brev) },
+                            { hentLagretInnholdVedlegg(brev) },
+                        ).innhold(),
+                    data = (
+                        migreringBrevDataService.opprettMigreringBrevdata(
+                            generellBrevData,
+                            automatiskMigreringRequest,
+                            brukerTokenInfo,
+                        ) as OmregnetBPNyttRegelverk
+                    ),
+                )
+        }
+    }
+
+    private suspend fun opprettBrevData(
+        brev: Brev,
+        generellBrevData: GenerellBrevData,
+        brukerTokenInfo: BrukerTokenInfo,
+        brevkode: BrevkodePar,
+    ): BrevData =
+        when (brev.prosessType) {
+            BrevProsessType.REDIGERBAR ->
+                brevDataMapper.brevDataFerdigstilling(
+                    generellBrevData,
+                    brukerTokenInfo,
+                    InnholdMedVedlegg({ hentLagretInnhold(brev) }, { hentLagretInnholdVedlegg(brev) }),
+                    brevkode,
+                )
+
+            BrevProsessType.AUTOMATISK -> brevDataMapper.brevData(generellBrevData, brukerTokenInfo)
+            BrevProsessType.MANUELL -> ManueltBrevData(hentLagretInnhold(brev))
+        }
+
+    private fun hentLagretInnhold(brev: Brev) =
+        requireNotNull(
+            db.hentBrevPayload(brev.id),
+        ) { "Fant ikke payload for brev ${brev.id}" }.elements
+
+    private fun hentLagretInnholdVedlegg(brev: Brev) =
+        requireNotNull(db.hentBrevPayloadVedlegg(brev.id)) {
+            "Fant ikke payloadvedlegg for brev ${brev.id}"
+        }
+}
